requirements_windows.txt
# Windows Perception Engine - Python Client Requirements

# Core dependencies (already in requirements.txt)
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
openai==1.3.0
requests==2.31.0

# ONNX Runtime for ML inference
onnxruntime>=1.17.0

# Computer vision
opencv-python==4.8.1.78
pillow>=10.0.0

# Audio processing
sounddevice==0.4.6
vosk==0.3.44

# Windows-specific
pywin32>=306
psutil>=5.9.0

# Transformers for FastVLM (optional, can use pure ONNX later)
transformers>=4.36.0

# NumPy (required by most ML libraries)
numpy>=1.24.0
